Make the style of body and header logo and links
I implemented a min-width to the body, so that the webpage will always
fill the screen.
I used flexbox to put the logo on the left and the links to the
right. I used padding to create space to the left and right, there
may be a better option.
I also styled the list and the links inside and used gap to create
space between the links.