Add style to third content section
The container is the box for the texts and the button.
I used a max-height on the container to center it vertically
inside the box while using flex:1;, there is a error with the place-
holder for the image, it goes off the page put I did not fix this yet.